- Fix some warnings spotted by MSVC - Define cmd_exit to ExitProcess if _DEBUG_MEM is not defined.   Fixes compilation, when _DEBUG_MEM is turned off.
svn path=/trunk/; revision=33545

+++ reactos/base/shell/cmd/cmdinput.c
@@ -518,7 +518,7 @@ VOID ReadCommand (LPTSTR str, INT maxlen)
 				else
 				{
 					LPCTSTR last = PeekHistory(-1);
-					if (last && charcount < _tcslen (last))
+					if (last && charcount < (INT)_tcslen (last))
 					{
 						PreviousChar = last[current];
 						ConOutChar(PreviousChar);
